AILoggerPlugin now checks whether the object at the usual Logs folder path is a file, and if it is, tries to dereference it as an alias. If it succeeds, it uses the destination of the alias as the new Logs folder.
This means that you can move your Logs folder to an encrypted disk image, and make an alias to the folder, and name that alias ?Logs? and put it into your Adium 2.0 folder, and Adium will use the folder on the encrypted disk image as your Logs folder. You don't even have to mount the disk image yourself?the Alias Manager will do that for you at the appropriate time.

Fixes #992. Refs #6925.