Instead of creating a window for every authorization requests, create a central "Authorization Requests" window. This has a few toolbar items: Authorize, Authorize & Add, and Deny. Their icons are currently the SSL image, because I suck at art and I needed a placeholder. Another commit is going to add get info buttons, deny and block, and a few other things.
This basically behaves the same, except you can mass-select and right-click/toolbar things away, right-click and get info on a person who is doing a request, and several nice features. This uses the dict stored in the array as the UI handle for libpurple, which may or may not be a terrible idea. In my super limited testing this works really well.

Fixes #11454. Fixes #9104. Fixes #6716 (there's a maximum height per row for the requests). Refs #9250. Refs #9767. Refs #7923.