Fix the range of the ?Custom Pitch? sliders. This fixes #12251, where any speech action with custom pitch had really high pitch.
The low end of the sliders was 90, which is just under twice the default value reported by Alex (45 and some change). All the system-supplied voices are in the 30?60 range. I wonder what units NSSpeechSynthesizer is using? In case you're wondering: The high end of the sliders was 300.

The new range is 0.01 to 90.