Fix moving tabs around causing crashes or behaving badly while "save open chats" is enabled. Fixes #12212.
This was caused by an error on my part when saved chats was added. We use uniqueContainerNumber to create unique containers, but we also restore these numerical values. When PSM says "hey, I need a new tab" we say "Sure! here, have this one" and give it an old tab.

This causes PSM to perform a wacky (in this situation) set of operations:
1) Add tab cell to new window.
2) Remove tab cell from old window. (triggers unbinding)
3) Set up bindings in new window.

However, the "add then remove" action causes it to be under-retained, lost, and confused. When a binding is generated, it's sent to a deallocated process (after the "remove"). When the new window and old window are not the same windows, however, PSM's actions are correct and proper.

The solution now is to not care what the created name of a chat is so saving is okay, but when we're creating a new window, we make sure it's new.