Change ICQ and AIM encryption settings to no/opportunistic/required encryption.
For ICQ, all accounts will be changed to opportunistic (as the SSL settings were wiped by 1.4.1).
For AIM, when SSL was on, encryption is now required. When SSL was off, encryption is now opportunistic.

Additionally, turn on clientLogin for AIM.

Fixes #15000
(transplanted from e9f94c290817c0c77fc71475826478218494665f)